+++ VidaForaneaServidor/API/resources/Place.py
@@ -2,8 +2,10 @@ from tabnanny import check
 from flask import request
 from flask_restful import Resource
 from http import HTTPStatus
+from flask_jwt_extended import get_jwt_identity, jwt_required
 
 from models.Place import Place, lista_places
+from models.Student import Student
 
 class ListPlaces(Resource):
 
@@ -23,6 +25,7 @@ class ListPlaces(Resource):
             })
         return {'data': data}, HTTPStatus.OK
 
+    @jwt_required
     def post(self):
         json_data = request.get_json()
 
@@ -39,7 +42,7 @@ class ListPlaces(Resource):
         imageNotEncodedToBytes = json_data.get('image')
         image = bytes(imageNotEncodedToBytes, 'utf-8')
         if Place.get_by_name(name):
-            return {'message': 'Lugar ya registrada'}, HTTPStatus.BAD_REQUEST
+            return {'message': 'Place already registered'}, HTTPStatus.BAD_REQUEST
 
         place = Place(
             name= name,
@@ -85,36 +88,64 @@ class ResourcePlace(Resource):
         }
         return data, HTTPStatus.OK
 
+    @jwt_required
     def delete(self, place_id):
         place = Place.get_by_id(place_id)
         if place is None:
             return {'message': 'Lugar no encontrado'}, HTTPStatus.NOT_FOUND
+        current_user = get_jwt_identity()
+        current_student = Student.get_by_enrollment(current_user)
+        if current_student.rol == 'estudiante' or current_user=='moderador' :
+            return {'message': 'Access is not allowed'}, HTTPStatus.FORBIDDEN
         place.delete()
         return  HTTPStatus.NO_CONTENT
 
+    @jwt_required
     def put(self, place_id):
         data = request.get_json()
-        place = next((place for place in lista_places if place.id == place_id ), None)
+        place = Place.get_by_id(place_id)
         if place is None:
             return {'message': 'Lugar no encontrado'}, HTTPStatus.NOT_FOUND
-
-        if place.status =="Pendiente":
-            place.status= 'Aprobado'
-
+        statusCheck = data['status']
+        if statusCheck == 1:
+           place.status= 'aprobado'
+        if statusCheck == 0:
+           place.status = 'pendiente'
         place.name = data['name']
         place.address = data['address']
         place.services = data['services']
         place.schedule = data['schedule']
         place.type_place = data['type_place']
-        place.image = data['image'].decode("utf-8")
+        place.image = bytes(data['image'], 'utf-8')
+        current_user = get_jwt_identity()
+        current_student = Student.get_by_enrollment(current_user)
+        if current_student.rol == 'estudiante' or current_user=='moderador' :
+            return {'message': 'Access is not allowed'}, HTTPStatus.FORBIDDEN
+        place.save()
         return data, HTTPStatus.OK
 
     def patch(self, place_id):
         place = next((place for place in lista_places if place.id == place_id ), None)
         if place is None:
             return {'message': 'Lugar no encontrado'}, HTTPStatus.NOT_FOUND
-        if place.status =='Pendiente':
-            place.status='Aprobado'
+        json_data = request.get_json()
+        name = json_data.get('name')
+        address = json_data.get('address')
+        services = json_data.get('services')
+        schedule = json_data.get('schedule')
+        type_place=json_data.get('type_place')
+        imageNotEncodedToBytes = json_data.get('image')
+        image = bytes(imageNotEncodedToBytes, 'utf-8')
+        place.name = name
+        place.address = address
+        place.services = services
+        place.schedule = schedule
+        if json_data.get('status') == 1:
+            status = "aprobado"
+        if json_data.get('status') == 0:
+            status = "pendiente"
+        place.type_place = type_place
+        place.image = image
 
         data = {
             'id': place.id,
